requirements.txt
# Core dependencies
docling>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0
python-multipart>=0.0.5

# Email validation
email-validator>=2.0.0

# Optional dependencies for enhanced functionality
python-dateutil>=2.8.0
typing-extensions>=4.0.0

# Logging and utilities
colorlog>=6.0.0

# Vector Store Dependencies
faiss-cpu>=1.7.0  # Use faiss-gpu if you have CUDA support
sentence-transformers>=2.2.0
torch>=1.9.0
numpy>=1.21.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
